Add tests for the documentation snippet
